fix: :bug: Fixes the connection issue in the connectionrequest listener caused by the incorrect validation of the number of peers in the room
