+++ lib/Room.js
@@ -4,6 +4,7 @@ const throttle = require('@sitespeed.io/throttle')
 const Logger = require('./Logger')
 const config = require('../config')
 const Bot = require('./Bot')
+const { sendNotification } = require('./handlerNotification');
 
 const logger = new Logger('Room')
 
@@ -187,8 +188,9 @@ class Room extends EventEmitter {
    * @param {Boolean} consume - Whether this peer wants to consume from others.
    * @param {protoo.WebSocketTransport} protooWebSocketTransport - The associated
    *   protoo WebSocket transport.
+   * @param {eventNotificationUri} - Uri to send notification webhook when peer close
    */
-  handleProtooConnection({ peerId, consume, protooWebSocketTransport }) {
+  handleProtooConnection({ peerId, consume, protooWebSocketTransport, eventNotificationUri }) {
    
     // Check if the room has reached the maxPeerCount limit
     if (this._protooRoom.peers.length >= this._maxPeerCount) {
@@ -248,7 +250,15 @@ class Room extends EventEmitter {
     peer.on('close', () => {
       if (this._closed) return
 
-      logger.debug('protoo Peer "close" event [peerId:%s]', peer.id)
+      logger.info('protoo Peer "close" event [peerId:%s]', peer.id)
+
+      // send notification webhook if eventNotificationUri is defined
+      const joinNotificationData = {
+          roomId: this._roomId,
+          peerId: peer.id,
+          event: 'peer-left',
+        };
+      sendNotification(eventNotificationUri, joinNotificationData);
 
       // If the Peer was joined, notify all Peers.
       if (peer.data.joined) {
