+++ lib/middlewares.js
@@ -3,8 +3,10 @@ const validateCreateRoomParams = (req, res, next) => {
   const { eventNotificationUri, maxPeerCount } = req.body;
 
   // Validate that eventNotificationUri is a string
-  if (!eventNotificationUri || typeof eventNotificationUri !== 'string') {
-    return res.status(400).json({ error: 'Invalid or missing eventNotificationUri' });
+  if (eventNotificationUri !== undefined && eventNotificationUri !== '') {
+   if (typeof eventNotificationUri !== 'string') {
+      return res.status(400).json({ error: 'Invalid eventNotificationUri, it must be a string if defined' });
+    }
   }
 
   // Validate that maxPeerCount is a number
